Refactor Gettext merging
Before this commit, we didn't know the Language type before creating a
Catalogue. As such, we were just blindly merged locale files with the
same name. This obviously has issues:

/locales/en/translation.mo
/locales/de/translation.mo

If we organized our translation files in the format above, Lens would
think that both `translation.mo`s are apart of the same locale and in
turn merge the results.

With this commit, the header information are now extracted immediately
after parsing allow for us to validate locales of the same type. If this
match fails, we won't merge them.

However, if a Gettext file hasn't defined the Language header then the
old merging method would still be utilized. This is a flaw that we can't
fix.