Skip plural-form check when translation isn't hash
Usually plural form syntax is in a structure like this:

apples:
  one: "I have %{count} apple"
  other: "I have %{count} apples"

However, there may be cases where the end-user wishes for only some items
to have plurals in a sequence. For instance:

byte:
  one:   "Byte"
  other: "Bytes"
kb: "KB"
mb: "MB"
gb: "GB"
tb: "TB"
pb: "PB"
eb: "EB"

When iterating, only byte should have a plural-form. Before this commit,
an error would be raised under this construction. This fixes that and
adds support for this sort of organization